Estimate diversity for community data

Estimation of diversity from community data. TODO

diversity_estimate(
  data_source,
  sel_method = c("taxonomic", "phylogenetic_diversity", "phylogenetic_nri",
    "phylogenetic_nti", "functional_rao", "functional_mpd", "functional_simpson"),
  round = TRUE,
  sample_size,
  rand = 999,
  iterations = 1000,
  abundance_weighted = TRUE,
  data_source_tree = NULL,
  data_source_traits = NULL
)

Arguments

data_source

Data frame with rows as levels and columns as taxa. First columns should be named sample_id.

sel_method

Character. Selected method of diversify estimation

"taxonomic" - individual-based Hill numbers of taxonomic diversity (diversity and evenness)

"phylogenetic_diversity" - phylogenetic Faith's diversity

"phylogenetic_nri" - phylogenetic Net Relatedness Index

"phylogenetic_nti" - phylogenetic Nearest Taxon Index

"functional_rao" - functional Rao's quadratic diversity

"functional_mpd" - functional Mean Pairwise Distance

"functional_simpson" - functional Simpsons Diversity Index

round

Logical. Should pollen values be rounded?

sample_size

Numeric. Only for "taxonomic". Minimum sample size

rand

Numeric. Number of randomisations

iterations

Numeric. Only for "phylogenetic". Number of iterations to use for each randomization (for independent swap and trial null models).

abundance_weighted

Logical. Only for "phylogenetic" or "functional". It FALSE, presence/absence data will be used for calculations.

data_source_tree

Only for "phylogenetic". Phylogenetic backbone tree constructed using ape package.

data_source_traits

Data frame. Row names be the same taxa as in data_source_community and vice versa.

Value

Data frame with diversity metric estimated for each level (sample). Possible outputs depending on the sel_method:

n0 - taxonomic Hill' N0

n1 - taxonomic Hill' N1

n2 - taxonomic Hill' N2

n1_minus_n2 - taxonomic evenness ratios (N1-N2)

n2_divided_by_n1 - taxonomic evenness ratios (N2/N1)

n1_divided_by_n0 - taxonomic evenness ratios (N1/N0)

n_taxa - Taxonomic richness

pd_faith - Faith's phylogenetic diversity

pd_mpd - phylogenetic Mean Pairwise Distance

pd_nri - phylogenetic Net Relatedness Index

pd_mntd - phylogenetic Mean Nearest Taxon Distance

pd_nti - phylogenetic Nearest Taxon Index

fd_rao - Rao's quadratic diversity

fd_mpd - Mean Pairwise Distance

fd_simpson - Simpsons Diversity Index

z_score - Standarise Effect Size (SES), calculated as (observed value - mean randomised value ) / sd of randomised value