Detect change points for all columns

Detect change points for all columns

get_change_points_all(
  data_source,
  var_name_order = "age",
  use_inbetween_values = TRUE,
  sel_output = c("only_inbetween", "both"),
  var_name_detection = NULL,
  direction = c("front", "back")
)

Arguments

data_source

Data.frame with columns to detect change points in

var_name_order

Character. Name of the columns, which should be use as a order of values

use_inbetween_values

Logical. If TRUE additional values will be placed in between values of var_name_order.

sel_output

Character. Only used for use_inbetween_values == TRUE Define what should the function output.

"only_inbetween" - only inbetween values,

"both" - both inbetween values and original data

var_name_detection

Character. Optional. Columns to use to change point detection will be automatically selected if they partly matched var_name_detection. For all columns use `NULL``.

direction

Which direction should the values be compared ("front" or "back")

Value

Data.frame with var_name_order and all columns detected based on var_name_detection, with 1 represent change points.